requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
